Updated BigLiquidStorage, BigGasStorage, BigBeautifulStorage and BigLockerStorage to have feature parity with base game
I was working with the automation, and noticed a missmatch in functionality with the base Liquide storage module.

By decompiling the source code I compared theLiquidReservoirConfig.cs to the BigLiquidStorageConfig.cs and added the missing components. 

Updated BigGasStorage to have feature parity with base game

Ditto to previous commit
Updated BigBeautifulStorageLocker to have feature parity with base game

Ditto to last
Updated BigStorageLocker to have feature parity with base game